Initial file refactor complete
