+++ backend/.env
@@ -1,2 +1,4 @@
 DATABASE_URL = postgresql://postgres:password123@localhost:5432/file_storage_service
-STORAGE_FILE_PATH = ../storage/
+FILE_STORAGE_PATH = source\storage
+HOST = 0.0.0.0
+PORT = 5433
